% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildRegTree.R
\name{BuildRegTree}
\alias{BuildRegTree}
\title{RerF Tree Generator}
\usage{
BuildRegTree(
  X,
  Y,
  FUN,
  paramList,
  min.parent,
  max.depth,
  bagging,
  replacement,
  store.oob,
  store.impurity,
  progress,
  rotate,
  honesty
)
}
\arguments{
\item{X}{an n-by-d numeric matrix (preferable) or data frame. The rows correspond to observations and columns correspond to features.}

\item{Y}{an array of length n containing the continuous responses to be regressed on.}

\item{FUN}{a function that creates the random projection matrix.}

\item{paramList}{parameters in a named list to be used by FUN. If left unchanged, default values will be populated, see \code{\link[rerf]{defaults}} for details.}

\item{min.parent}{the minimum splittable node size.  A node size < min.parent will be a leaf node. (min.parent = 6)}

\item{max.depth}{the longest allowable distance from the root of a tree to a leaf node (i.e. the maximum allowed height for a tree).  If max.depth=0, the tree will be allowed to grow without bound.  (max.depth=0)}

\item{bagging}{a non-zero value means a random sample of X will be used during tree creation.  If replacement = FALSE the bagging value determines the percentage of samples to leave out-of-bag.  If replacement = TRUE the non-zero bagging value is ignored. (bagging=.2)}

\item{replacement}{if TRUE then n samples are chosen, with replacement, from X. (replacement=FALSE)}

\item{store.oob}{if TRUE then the samples omitted during the creation of a tree are stored as part of the tree.  This is required to run OOBPredict(). (store.oob=FALSE)}

\item{store.impurity}{if TRUE then the increase in information gain is stored for every split. This is required to run FeatureImportance() (store.impurity=FALSE)}

\item{progress}{if true a pipe is printed after each tree is created.  This is useful for large datasets. (progress=FALSE)}

\item{rotate}{if TRUE then the data matrix X is uniformly randomly rotated. (rotate=FALSE)}

\item{honesty}{if TRUE then OOB samples will be used for local leaf node estimates.}
}
\value{
Tree
}
\description{
Creates a single decision tree based on an input matrix and similarity matrix. This is the function used by rerf to generate trees.
}
