% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrerfHelperFunctions.R
\name{checkInputMatrix}
\alias{checkInputMatrix}
\title{Determine if given input can be processed by Urerf.}
\usage{
checkInputMatrix(X)
}
\arguments{
\item{X}{an Nxd matrix or Data frame of numeric values.}
}
\value{
stops function execution and outputs error if invalid input is detected.
}
\description{
Determine if given input can be processed by Urerf.
}
