% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{flipWeights}
\alias{flipWeights}
\title{Change the sign of the weights}
\usage{
flipWeights(x)
}
\arguments{
\item{x}{a list of unique.projections from the intermediate steps of
the FeatureImportance function.}
}
\value{
x with sign of weights flipped.
}
\description{
A helper function to extract the feature weights from the projection
vector stored in a tree object. Used in
\code{RunFeatureImportanceBinary}.
}
