% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpPredict.R
\name{fpPredict}
\alias{fpPredict}
\title{Packs a forest and saves modified forest to disk for use by PackPredict function}
\usage{
fpPredict(forest = NULL, X = NULL)
}
\arguments{
\item{forest}{forest data structure returned from fpRerF function.}

\item{X}{an n by d numeric matrix (preferable) or data frame used to train the forest.}
}
\description{
Efficiently packs a forest trained with the RF option.  Two intermediate data structures are written to disk, forestPackTempFile.csv and traversalPackTempFile.csv.  The size of these data structures is proportional to a trained forest and training data respectively.  Both data structures are removed at the end of the operation.  The resulting forest is saved as forest.out.  The size of this file is similar to the size of the trained forest.
}
\examples{

library(rerf)
forest <- fpRerF(as.matrix(iris[, 1:4]), as.numeric(iris[[5L]])-1)
predictions <- fpPredict(forest, as.matrix(iris[, 1:4]))
}
