% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpRerF.R
\name{fpRerF}
\alias{fpRerF}
\title{Packs a forest and saves modified forest to disk for use by PackPredict function}
\usage{
fpRerF(X = NULL, Y = NULL, csvFileName = NULL, columnWithY = NULL,
  maxDepth = Inf, minParent = 1, numTreesInForest = 500,
  numCores = 1, numTreeBins = NULL, forestType = "binnedBaseRerF",
  nodeSizeToBin = NULL, nodeSizeBin = NULL, mtry = NULL,
  mtryMult = NULL, seed = sample(1:1e+06, 1))
}
\arguments{
\item{X}{an n by d numeric matrix (preferable) or data frame used to train the forest.}

\item{Y}{a numeric vector of size n.  If the Y vector used to train the forest was not of type numeric then a simple call to as.numeric(Y) will suffice as input.}

\item{csvFileName}{the name of a headerless csv file containing combined data and labels.}

\item{columnWithY}{is the column in the headerless csv file containing class lables.}

\item{maxDepth}{int the maximum allowed tree height/depth (path distance between root and leaves). (maxDepth = Inf, i.e. largest system int)}

\item{minParent}{is the size of nodes that will not be split (minParent=1)}

\item{numTreesInForest}{the number of trees to grow in the forest (numTreesInForest=500)}

\item{numCores}{is the number of cores to use when training and predicting with the forest (numCores=1)}

\item{numTreeBins}{the number of bins to store the forest.  Each bin will contain numTreesInForest/numTreeBins trees.  Only used when forestType=="binned*" (numTreeBins= numCores)}

\item{forestType}{the type of forest to grow: binnedBase, binnedBaseRerF (forestType="binnedBaseRerF")}

\item{mtry}{the number of features to consider when splitting a node (mtry=ncol(X)^.5)}

\item{mtryMult}{the average number of features combined to form a new feature when using RerF (mtryMult=1)}

\item{NodeSizeToBin}{the minimum node size to use stratified subsampling (NodeSizeToBin=NULL)}

\item{NodeSizeBin}{the size of the stratified subsample chosen when NodeSizeToBin criteria is met (NodeSizeBin=NULL)}

\item{forestType}{the type of forest to grow: binnedBase, binnedBaseRerF, rfBase, rerf (forestType="binnedBaseRerF")}
}
\description{
Efficiently packs a forest trained with the RF option.  Two intermediate data structures are written to disk, forestPackTempFile.csv and traversalPackTempFile.csv.  The size of these data structures is proportional to a trained forest and training data respectively.  Both data structures are removed at the end of the operation.  The resulting forest is saved as forest.out.  The size of this file is similar to the size of the trained forest.
}
\examples{
library(rerf)
## setup data
X <- as.matrix(iris[, 1:4])
Y <- as.numeric(iris[[5]]) - 1
forest <- fpRerF(X, Y, numCores = 2L)
(training.error <- mean(fpPredict(forest, X) != Y))

}
