% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{getWeights}
\alias{getWeights}
\title{Extract feature weights from the sparse projection vector.}
\usage{
getWeights(x)
}
\arguments{
\item{x}{a list of unique.projections from the intermediate steps of
the FeatureImportance function.}
}
\value{
list of unique feature weights
}
\description{
A helper function to extract the feature weights from the projection
vector stored in a tree object.
}
