% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist-data.R
\docType{data}
\name{mnist}
\alias{mnist}
\title{A subset of the MNIST dataset for handwritten digit classification}
\format{A list with four items: Xtrain is a training set matrix with 6000 rows (samples) and 784 columns (features),
Xtrain is an integer array of corresponding training class labels, Xtest is a test set matrix of 10000 rows and 784
columns, and Ytest is the corresponding class labels. Rows in Xtrain and Xtest correspond to different images of digits,
and columns correspond to the pixel intensities in each image, obtained by flattening the image pixels in column-major ordering.}
\source{
\href{http://yann.lecun.com/exdb/mnist/}{MNIST}
}
\usage{
data(mnist)
}
\description{
A dataset consiting of 10 percent of the MNIST training set and the
full test set
}
\examples{
data(mnist)
}
\references{
Y. LeCun, L. Bottou, Y. Bengio, and P. Haffner. "Gradient-based learning applied to document recognition."
Proceedings of the IEEE, 86(11):2278-2324, November 1998.
}
\keyword{datasets}
