% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScaleNode.R
\name{PredictScale01}
\alias{PredictScale01}
\title{Scale new data in current Node to [0,1] based on teh scaling factors
from training.}
\usage{
PredictScale01(Xnode, scalingFactors)
}
\arguments{
\item{Xnode}{an n by d numeric matrix or data frame containing the data assigned to the current node. The rows correspond to observations and columns correspond to features.}

\item{scalingFactors}{a list with names elements \code{min} and
\code{diff} containing vectors of the minima and difference in maxima
and minima across columns of Xnode.}
}
\value{
the scaled data as a matrix
}
\description{
Scales the data in the current node to live in [0,1] over the columns
with the scaling factors used when training the forest.
}
