% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerf-package.R
\docType{package}
\name{rerf-package}
\alias{rerf}
\alias{rerf-package}
\title{rerf-package}
\description{
The rerf package provides an R implementation of the randomer forest
algorithm developed by Tomita et al.(2016) <arXiv:1506.03410v2>,
along with variants such as Structured Randomer Forest (S-RerF) and
Similarity Randomer Forest (SmerF).
}
\section{The main function \code{RerF}}{

The \code{RerF} function is the primary access point to the
algorithm. Variants of the algorithm can be controlled by specifying
the appropriate set of parameters, explained in the sections below.
}

\section{RandMat*}{

The set of \code{RandMat*} functions specify the distribution of
projection weights used to split the data at each node.

\describe{
  \item{RandMatBinary}{samples projection weights from \eqn{{-1,1}}.}
  \item{RandMatContinuous}{samples projection weights from \eqn{N(0,1)}}
  \item{RandMatPoisson}{samples projection weights from a \eqn{N(0,1)} with sparsity \eqn{~ Poisson(\lambda)}}
}
}

\section{RerF variants}{

\describe{
  \item{RF}{Use \code{FUN = RandMatRF} in the call to \code{RerF}}
  \item{RerF}{Use \code{FUN = RandMatBinary} in the call to \code{RerF}}
  \item{S-RerF}{Use \code{FUN = RandMatImagePatch} in the call to \code{RerF}}
  \item{SmerF}{Set \code{task = "similarity"} in the call to \code{RerF}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/neurodata/R-RerF}
  \item Report bugs at \url{https://github.com/neurodata/R-RerF/issues}
}

}
\author{
\strong{Maintainer}: Meghana Madhyastha \email{mmadhya1@jhu.edu} [contributor]

Authors:
\itemize{
  \item James Browne \email{jbrowne6@jhu.edu}
  \item Tyler Tomita \email{ttomita2@jhmi.edu}
}

Other contributors:
\itemize{
  \item Jaewon Chung \email{j1c@jhu.edu} [contributor]
  \item Jesse Patsolic \email{jpatsol1@jhu.edu} [contributor]
  \item Benjamin Falk \email{falk.ben@jhu.edu} [contributor]
  \item Joshua Vogelstein \email{jovo@jhu.edu} [thesis advisor]
}

}
