% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPredictLeaf.R
\name{RunPredictLeaf}
\alias{RunPredictLeaf}
\title{Calculate similarity using a single tree.}
\usage{
RunPredictLeaf(X, tree)
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame which was used to train the provided forest.}

\item{tree}{a tree from a forest returned by RerF.}
}
\value{
similarity based on one tree
}
\description{
This is the base function called by ComputeSimilarity.
}
