% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPredictReg.R
\name{RunPredictReg}
\alias{RunPredictReg}
\title{Predict class labels on a test set using a single tree.}
\usage{
RunPredictReg(X, tree)
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame which was used to train the provided forest.}

\item{tree}{a tree from a forest returned by RerF.}
}
\value{
predictions an n length vector of prediction based on the tree provided to this function
}
\description{
This is the base function called by Predict.
}
