% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPredictSim.R
\name{RunPredictSim}
\alias{RunPredictSim}
\title{Predict class labels on a test set using a single tree.}
\usage{
RunPredictSim(X, Xtrain, tree, OOB)
}
\arguments{
\item{X}{an n sample by d feature matrix (preferable) or data frame which was used to train the provided forest.}

\item{Xtrain}{an n by d numeric matrix (preferable) or data frame. This should be the same data matrix/frame used to train the forest. If provided, then the output will be an ntest-by-ntrain matrix of similarities between training and test points (Xtrain=NULL)}

\item{tree}{a tree from a forest returned by RerF.}

\item{OOB}{if TRUE and Xtrain is provided, then the tree predicts only the similarities of the out-of-bag training points to the test points. (OOB=TRUE)}
}
\value{
predictions an n length vector of prediction based on the tree provided to this function
}
\description{
This is the base function called by Predict.
}
