% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UrerfHelperFunctions.R
\name{TwoMeansCut}
\alias{TwoMeansCut}
\title{Find minimizing Two Means Cut for Vector}
\usage{
TwoMeansCut(X)
}
\arguments{
\item{X}{a one dimensional vector}
}
\value{
list containing minimizing cut point and corresponding sum of left and right variances.
}
\description{
Find minimizing Two Means Cut for Vector
}
